move openapiadapter to apiadapter
